fix(gguf): enforce version-appropriate KV key lengths

+++ llm-runner/tests/dispatch_integration.rs
@@ -181,7 +181,7 @@ fn make_test_gguf(path: &PathBuf) {
     buf.extend_from_slice(&(kv_pairs.len() as u64).to_le_bytes());
     for kv in &kv_pairs {
         let key_bytes = kv.key.as_bytes();
-        buf.extend_from_slice(&(key_bytes.len() as u32).to_le_bytes());
+        buf.extend_from_slice(&(key_bytes.len() as u64).to_le_bytes());
         buf.extend_from_slice(key_bytes);
         buf.extend_from_slice(&kv.value_type.to_u32().to_le_bytes());
         write_kv_value(&mut buf, &kv.value);
